06-seccion_tool_connect_call/langchain-mcp-adapters/pyproject.toml
[project]
name = "langchain-mcp-adapters-pct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]
